pyproject.toml
[project]
name = "monitoring-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "dependency-injector (>=4.48.2,<5.0.0)",
    "fastapi-pagination (>=0.15.0,<0.16.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
